pyproject.toml
[tool.poetry]
name = "ncgt-patterns"
version = "0.1.0"
description = ""
authors = ["Jordan Stewart <jstewart@newscorp.com>"]
readme = "README.md"
packages = [{include = "strategy_pattern"}]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.2"
toml = "^0.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"